LayoutTests/js/dfg-make-rope-all-empty-strings-expected.txt

Check that the DFG can handle MakeRope or ValueAdd on all empty strings.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS bar() is ""
PASS successfullyParsed is true

TEST COMPLETE
